pyproject.toml
[project]
name = "uv-lens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "packaging>=24",
    "pyyaml>=6",
    "rich>=13",
    "textual>=0.60",
    "tomlkit>=0.13",
]

[project.scripts]
uv-lens = "uv_lens.cli:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.23",
]

[build-system]
requires = ["uv_build>=0.9.15,<0.10.0"]
build-backend = "uv_build"
